修复toSring()

+++ miscroservice_user/src/main/java/com/e_commerce/miscroservice/csq_proj/service/CsqCollectionService.java
@@ -9,7 +9,7 @@ public interface CsqCollectionService {
 	 * @param serviceId
 	 * @param userId
 	 */
-	Object clickCollection(Long serviceId, Integer userId);
+	Object clickCollection(Long serviceId, Long userId);
 
 	/**
 	 * 收藏列表
